Render in stateChange
Back in March I got the idea that we should
render the page before updating the URL
because of timing and errors. Now after
chasing bugs for a while I've decided we
should do it they way it was intended and
deal with errors some other way.

I think this code now handles URL changes
properly in modern and old (IE) browsers.