Attempt at issue #25 retry host rendering when client fails
This is one approach to the issue of not knowing a URL is not a book until after we've asked
for it. I implemented a retry to the host when client-side rendering fails. I'm not convinced
this is the best solution so I'm keeping it in a branch for now.