Issue #96, EPUB.
Now you can download a book in EPUB format. They appear to work on the iPad. They
don't look so good in the Ubuntu e-reader but maybe there is some problem with it?

I have done almost nothing to the style. It will be easy to tweak the stylesheet
in EPub/page.html if needed.