A first hack at eliminating media queries using javascript.
I bet this can be done better. It was just the first thing I tried. It
seems to work fine but is a bit jerky, especially on my old 1st generation
iPad when rotating.

Of course it doesn't include all the neat hacks to tweak the buttons
and such on small displays. I imagine they would be triggered by the
presence of the "tiny" class on the body.