Adding JavaScript for installing net-score.
This tag will load a network performance test into an invisible iframe
to test network performance in the background.  These tests are
throttled based on a /24 for IPv4 and /64 for IPv6) every 60 seconds.
The iframe is loaded after the entirety of the parent page has rendered
and should not affect the performance or user experience for the parent
page.

Documentation for net-score and the tests that are run can be found at
http://info.net-score.org.