Updating CSS to use display over visibility.
The net-score iframe was still taking up space in the TarHeelReader DOM
and this caused problems with the layout as elements were appended to
the DOM.  By replacing visibility:hidden with display:none the net-score
iframe no longer consumes space on the rendered page.  This means that
as the user progresses through the application the layout is not changed
by the presence of the net-score iframe.  Tested on:

Windows XP, with IE7, IE8, FF 24, Chrome 29, Opera 16
Windows 7 and 8 with IE7, IE8, IE9, IE10, FF 24, Chrome 29, Opera 16
Ubuntu 12.04 with FF 24, Chrome 30, Opera 16
OS X 10.6.8 with Safari 5, FF 25, Chrome 29, Opera 16
Android 4.4.8 with Chrome 29